requirements.txt
python-binance==1.0.19
python-dotenv==1.0.1
requests>=2.31.0
rich==13.7.1
reportlab==4.2.5
